requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
reportlab>=4.0.0
gunicorn>=22.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
